+++ crates/client/sync/src/l2.rs
@@ -49,7 +49,7 @@ pub struct L2StateUpdate {
 pub struct L2VerifyApplyConfig {
     block_import: Arc<BlockImporter>,
     backup_every_n_blocks: Option<u64>,
-    flush_every_n_blocks: u16,
+    flush_every_n_blocks: u64,
     stop_on_sync: bool,
     telemetry: TelemetryHandle,
     validation: BlockValidationContext,
@@ -77,7 +77,7 @@ async fn l2_verify_and_apply_task(
     while let Some(block) = channel_wait_or_graceful_shutdown(pin!(block_conv_receiver.recv()), &ctx).await {
         let BlockImportResult { header, block_hash } = block_import.verify_apply(block, validation.clone()).await?;
 
-        if header.block_number - last_block_n >= flush_every_n_blocks as u64 {
+        if header.block_number - last_block_n >= flush_every_n_blocks {
             last_block_n = header.block_number;
             backend.flush().context("Flushing database")?;
         }
@@ -224,7 +224,7 @@ pub struct L2SyncConfig {
     pub verify: bool,
     pub sync_polling_interval: Option<Duration>,
     pub backup_every_n_blocks: Option<u64>,
-    pub flush_every_n_blocks: u16,
+    pub flush_every_n_blocks: u64,
     pub pending_block_poll_interval: Duration,
     pub ignore_block_order: bool,
     pub warp_update: bool,
